+++ src/components/Headers/Header.js
@@ -9,7 +9,7 @@ import Icon from 'react-native-vector-icons/FontAwesome';
 import styles from "./HeaderStyles";
 import { stackNavigator, withNavigation } from "react-navigation";
 import ColorConstants from "../../assets/ColorConstants";
-import ComponentTest from "../../screens/FramedScreens/ComponentTest";
+import ComponentTest from "../../screens/FramedScreens/RegAsset_1_Draft_1";
 const bgImage = require("../../assets/main-bg.png")
